[TDP-485] Fix Ceil with big number

+++ dataprep-backend-common/src/main/java/org/talend/dataprep/transformation/api/action/metadata/math/Floor.java
@@ -39,8 +39,8 @@ public class Floor extends AbstractMath
     }
 
     @Override
-    protected int compute(double from) {
-        return (int) Math.floor(from);
+    protected long compute(double from) {
+        return (long) Math.floor(from);
     }
 
 }
